refactor(App): extract useQueriesPanel composable
Moves the role-specific quick-queries footer state (queriesExpanded,
queriesListRef, auto-scroll watch, onQueryClick) into a dedicated
composable. The composable takes userInput / textareaRef / sendMessage
as explicit deps so its signature documents exactly what it needs
from the host component.
